+++ changelog.md
@@ -5,6 +5,7 @@ Codion Change Log
 ### is.codion.common.core
 - Value.Notify.WHEN_CHANGED and WHEN_SET renamed CHANGED and SET respectively.
 - MethodLogger.Entry.getChildren() renamed children(), enterMessage() renamed message(), ArgumentToString renamed ArgumentFormatter.
+- PropertyValue.propertyName() renamed name().
 ### is.codion.common.db
 - Database.URL_SCOPED_INSTANCE configuration value added.
 ### is.codion.swing.common.model